feat(kmip-3.0): fix OAEP crypto params, Link attributes, credential/PKCS11 payloads, and misc KMIP 3.0 gaps
Continues closing KmipV30VerificationTest gaps. KmipV21VerificationTest stays
1410/1410 throughout (hard regression guard). KmipV30VerificationTest:
mandatory 1200 total, up from 1145 passing to 1169 passing; optional 34
total, up from 29 passing to 32 passing.

CryptographicParameters OAEP fields:
- Added SaltLength (Integer), PSource (ByteString), MaskGenerator (wired in,
  class already existed), and new MaskGeneratorHashingAlgorithm (Enumeration
  reusing HashingAlgorithm's value domain under its own tag - verified no
  distinct value table exists for it in the spec).
- Root cause for the whole OAEP test cluster was smaller than the missing
  fields alone: MaskGenerator's own Standard.MFG1 description was the typo
  "Mfg1" instead of the wire value "MGF1", silently dropping every OAEP
  request payload.

UniqueIdentifier-family (KMIP 3.0 §4.68 "Identifier" wire encoding):
- EncodingType gains IDENTIFIER/REFERENCE/NAME_REFERENCE (bytes 0x0C-0x0E,
  confirmed present in the v3.0 Item Type Enumeration table and absent from
  v2.1's, which stops at 0x0B).
- UniqueIdentifier (core, unforked - used across hundreds of cross-version
  shared structures, and shown by the v3.0 corpus to vary encoding *within*
  a single version) extends its existing sourceEncoding field/registration
  loop to also accept Identifier/Reference/NameReference, alongside its
  pre-existing TextString/Integer support.
- PrivateKeyUniqueIdentifier/PublicKeyUniqueIdentifier (core) are monomorphic
  per version (always TextString pre-3.0, always Identifier in 3.0), so
  narrowed to V1_2-V2_1 and forked into model.v3_0.type Identifier-only
  siblings instead, along with the payloads that use them (CreateKeyPair
  response, ReKeyKeyPair request/response).
- Fixed a latent registry collision this surfaced: core UniqueIdentifier and
  v2_1.enumeration.UniqueIdentifier (the pre-existing Enumeration variant,
  e.g. "IDPlaceholder") were both registering for the same
  (tag, ENUMERATION) key; scoped core's registration to exclude ENUMERATION,
  which that class already owns exclusively. Also fixed a spec-compliance
  typo this collision had masked: "IdPlaceholder" -> "IDPlaceholder".

Message envelope (RequestHeader/ResponseHeader/RequestBatchItem/ResponseBatchItem):
- v2.1 classes narrowed to drop V3_0; new model.v3_0.structure forks add/drop
  the fields that actually changed (Server Hashed Password on ResponseHeader;
  Batch Order Option and Unique Batch Item ID dropped).
- ServerHashedPassword moved from model.v2_1.type to model.v3_0.type (it's
  v3.0-only, was incorrectly scoped to also claim V2_1).

ModifyAttribute / DeleteAttribute (v2.1+v3.0 fork):
- v1.2's ModifyAttributeOpResponsePayload incorrectly claimed V2_1/V3_0 with
  the wrong (index-based Attribute) shape, causing an NPE; narrowed it.
- DeleteAttributeOpRequestPayload/ResponsePayload had no v2.1/v3.0 fork at
  all. Both forked to model.v2_1.structure.{request,response}.payload,
  mirroring the already-correct AddAttribute fork (UniqueIdentifier +
  NewAttribute request, UniqueIdentifier-only response for ModifyAttribute;
  UniqueIdentifier + AttributeReference request for DeleteAttribute).

CreateCredential / PKCS11 (v3.0):
- CreateCredentialOpRequestPayload was missing a required Attributes field
  (spec Table 274) and its deserializers only handled a wrapped
  <CredentialValue> tag when the v3.0 corpus emits the credential body
  unwrapped (<PasswordCredential>, <Certificate>, etc. as direct children).
  Fixed both; made PasswordCredential/HashedPasswordCredential/OtpCredential/
  Certificate implement CredentialValue directly.
- New CredentialLink attribute (tag existed, no class) used by CreateUser.
- Pkcs11OpRequestPayload/ResponsePayload were missing CorrelationValue,
  used to thread a PKCS#11 session across C_Initialize/C_GetInfo/C_Finalize
  calls.
- Operation.PKCS_11's description was "Pkcs11"; every occurrence in the
  actual OASIS corpus (v2.1 and v3.0 alike) uses "PKCS_11" - fixed.
- Pkcs11Function's Standard enum remains an intentional empty stub:
  KMIP 3.0 spec 11.39 states its values are "the 1-based offset count of the
  function in the CK_FUNCTION_LIST_3_0 structure as specified in [PKCS#11]",
  an external standard not present in this repo. Left unimplemented rather
  than guessing values - PKCS11-M-1-30 stays failing pending that data.

Other attribute/tag fixes (each verified missing/incorrect against the raw
v3.0 spec HTML before fixing, and cross-checked that no v2.1 corpus file
exercises the changed tag where relevant, so KmipV21VerificationTest
couldn't have caught them):
- ObjectClass (new in 3.0, tag 0x42019E) existed but was never registered as
  a KmipAttribute, so it could never appear inside an Attributes structure;
  same bug class as ProtectionLevel fixed earlier this effort.
- OpaqueDataType has no standard values at all (vendor-extension only, per
  spec 11.35's empty table); the corpus writes an unregistered extension
  directly as a hex literal (type="Enumeration" value="0x80000001") rather
  than a pre-registered description - fromName now auto-registers hex
  literals for this class.
- NextLink/PreviousLink/PrivateKeyLink/PublicKeyLink were modeled as
  structures wrapping a nested UniqueIdentifier and were never registered as
  KmipAttribute; spec 4.35.9/14/15/16 defines them as a scalar Reference
  directly on the attribute's own tag. Redesigned as scalar
  KmipDataType+KmipAttribute classes.
- New GroupLink class (tag existed, encoding NameReference, no class).
- KmipTag RANDOM_IV's wire description was "RandomIv"; every corpus
  occurrence (v2.1 and v3.0) uses "RandomIV" - fixed. No v2.1 test case
  exercises this tag, so this couldn't regress KmipV21VerificationTest.
- CheckOpRequestPayload (v1.2, unforked) was scoped only to V1_2 while its
  response sibling already claimed through V3_0 and its shape is identical
  across versions - extended to match, no fork needed.
- RngSeedOpRequestPayload (v1.2, unforked) similarly under-scoped vs. its
  response sibling despite an identical DataByteString shape across
  versions - extended.
- KmipV30VerificationTest's placeholder substitution was missing
  $DIGEST_VALUE/$AUTHENTICATED_ENCRYPTION_TAG/$IV_COUNTER_NONCE/$MAC_DATA,
  causing hex-parse crashes that silently dropped whole batch items.

Known remaining gaps, left deliberately unfixed (documented, not chased):
- CryptographicUsageMask flag ordering: the OASIS v3.0 corpus itself is
  internally inconsistent (both "Decrypt Encrypt" and "Encrypt Decrypt" for
  the identical bitmask, occasionally in the same file) - no canonical
  choice satisfies both, and our ascending-bit-value ordering is a
  defensible choice already matching the majority pattern.
- Hex-case on one marker byte in a few SignatureVerify/Encrypt payloads
  (corpus writes a deliberate uppercase "FF" prefix, we emit lowercase) -
  non-semantic, same byte value either way.
- PKCS11-M-1-30, per the Pkcs11Function note above.

+++ src/main/java/org/purplebean/kmip/codec/ttlv/deserializer/model/v2x1/type/ProtectionPeriodTtlvDeserializer.java
@@ -28,7 +28,7 @@ public class ProtectionPeriodTtlvDeserializer extends
   @Override
   protected void setValue(ProtectionPeriod.ProtectionPeriodBuilder builder, byte[] tag, byte type,
                           ByteBuffer p, TtlvMapper mapper) throws IOException {
-    builder.value(mapper.readValue(p, Long.class));
+    builder.value(mapper.readValue(p, Integer.class));
   }
 
   @Override
